FAPI2: Go back to direct multicast group mapping
The indirect multicast group mapping table introduced in P10 ended up
causing more harm than good, so let's move back to a direct mapping of
multicast groups.

We still need more multicast groups than we have group IDs available,
so we are defining only very few static groups that a procedure can assume
to work out of the box, and HWPs are now required to set up multicast as
needed before they begin work. Any group that is not defined statically is
available to them for dynamic setup.

A separate utility function will be provided (not part of the base FAPI2
API) to take care of MC group setup.

Change-Id: Ieb20b8e1acb359284553e60fe8a621262c7d12f0
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127452
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Kahn C Evans <kahnevan@us.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/128340
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>