archive: Add PowerPC filter for better code compression
Enhance tinflate by adding a new compression method, METHOD_DEFLATEPPC,
which adds a PowerPC branch transformation step before compressing to
improve redundancy and therefore compression ratio.

Note that this method is slower than pure Deflate since we need to undo
the transformation after decompression, and it might in fact cause worse
compression ratio if used on a file that isn't mostly PowerPC code. So
it is only suitable for very specific files, but on those it helps a lot.

Change-Id: Iaa7b6adb6ed8ae1c9af0f6605ecea186fcf617a9
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127605
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Doug Gilbert <dgilbert@us.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/127609
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>