Add new TargetState Enums
Add two new TargetState Enums: TARGET_STATE_NOT_PRESENT and
TARGET_STATE_ALL. TARGET_STATE_NOT_PRESENT represents targets
that are not present in the system. This enum is not valid in
eCMD/Cronus and is currently unused. TARGET_STATE_ALL represents
targets that are both present and non-present.

Change-Id: I37a145ae823cf852699bbbfab8108575d9f7ea57
RTC: 287397
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/129426
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Steven B Janssen <janssens@us.ibm.com>
Reviewed-by: Kahn C Evans <kahnevan@us.ibm.com>
Reviewed-by: NAREN AMMANDA DEVAIAH <naren.devaiah@in.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/129428
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>