Splitting the IRQ list and routing table into a TCC and OCC part
Note that minimal changes to files was the goal. The only defines that
get a name change is the interrupt names themselves. So all interrupt
names for the TCC get the name, TCCHW_IRQ_<xyz>. No changes for the OCC.

Note that all other OCCHW_<abc> defines remain unchanged for the TCC.
This is because it spins out of control with "domino effect" changes once
we start changing these to TCCHW_<abc> because most of the OCCHW_<abc>
defines are used in other iota and pk files.

- TCC part done.
- Updated other "application" makefiles for pgpe, pce, iota and pk tests.
- OCC part done.

Change-Id: If5b6711cd6a2738ed1235e525c59efb1cbdf4f03
Original-Change-Id: Id30505bf645887b51b64fc71ed40bf0a3b571e99
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/129983
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Prem Shanker Jha <premjha2@in.ibm.com>
Reviewed-by: RANGANATHPRASAD G. BRAHMASAMUDRA <prasadbgr@in.ibm.com>
Reviewed-by: Gregory S Still <stillgs@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/135183
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>