archive: Split archive_config.H into per-platform files
Do not provide a default config header, instead require each platform
that uses the archive code to provide its own config. This way we don't
have to try (and fail) to provide a one-size-fits-all config at a central
location.

Include a config for iota, one for the archive test suite and an example
config as a means of documentation.

Change-Id: Ie1e027a86cb6b71d6b74cc101bf0a5abf00ead04
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131409
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>