archive: Split archive_config.H into per-platform files
Do not provide a default config header, instead require each platform
that uses the archive code to provide its own config. This way we don't
have to try (and fail) to provide a one-size-fits-all config at a central
location.

Include a config for iota, one for the archive test suite and an example
config as a means of documentation.

Change-Id: I3db9a5b7ce1d0bf577ca2ee22d367fef0085caa2
Original-Change-Id: Ie1e027a86cb6b71d6b74cc101bf0a5abf00ead04
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131579
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Doug Gilbert <dgilbert@us.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131732
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>