paktool list embedded pak expansion, output cleanup
paktool list now has a -e/--expand option to list the contents
of paks embedded in pak.  This is a recursive search.

The output object is now a proper singleton implementation that
self declares just by importing the module.  This will make it
easier for other projects to consume as a module.

Change-Id: I9436e3046db54b94a0a13c20f65eb632d1c1ea48
Signed-off-by: Jason Albert <albertj@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131920
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/133509
Reviewed-by: Team <hostboot@us.ibm.com>
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>