archive: Add template code for concatenating archives
Add a code template for others to copy who wish to append an archive
to another. Also add an append_archive method to FileArchive which
uses said template to provide archive append functionality.

Rename the append method to append_file for clarity.

Update the test suite to test file and archive appending, split out
the append tests from the decompression tests in the process.

Change-Id: I5a58f3a7e165f43429cb6007925237c5e5b6cda9
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/133119
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Kahn C Evans <kahnevan@us.ibm.com>
Reviewed-by: Doug Gilbert <dgilbert@us.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/133122
Reviewed-by: Team <hostboot@us.ibm.com>
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>