cmdtable: Calibrate poll timeout, add infinite timeout option
In the POLL instruction, add the option to poll indefinitely if the
error code parameter is zero.

Also update the poll delay loop so that the timeout on a 600 MHz
PPE is about 100ms. Since the timeout is hardcoded in ROM,
prefer a long timeout that will take a long time to detect an error
over a short timeout which might abort prematurely if hardware is
taking longer than expected.

Change-Id: I9cf4cf2374d55b920bb1553fe7382671bb8fd70e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/134400
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Rajees P P <rajerpp1@in.ibm.com>
Reviewed-by: Akhilesh S <akhis023@in.ibm.com>
Reviewed-by: AMIT J TENDOLKAR <amit.tendolkar@in.ibm.com>