"OCCHW" renaming (part 2)
This commit extends the renaming of "OCCHW" to "PMHW" whereever this
makes sense, as follows.

"OCCHW" define names that could potentially be used in Iota based
PPE images on both platforms have been renamed to "PMHW".

"OCCHW" define names that are used only on the Spinal side, including
those that are exclusive to the 405 and its Gpes, keep their "OCCHW"
names.

"OCCHW" defines names that are used only on the Taps were renamed to
"TCCHW" in the original renaming commit 134295.

This commit does not rename any of the "occhw_xyz" functions
nor filenames. (This was only done in the original commit 134295.)

Renaming progress:
- APPCFG_OCC_INSTANCE_ID -> APPCFG_PM_INSTANCE_ID
- OCCHW_INST_ID_xyz      -> PMHW_INST_ID_xyz
- OCCHW_<panic_codes>    -> PMHW_<panic_codes>

Change-Id: Idefac562c38eec98d0299f180fe9eca68f122d7e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/135173
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: RANGANATHPRASAD G. BRAHMASAMUDRA <prasadbgr@in.ibm.com>
Reviewed-by: Prem Shanker Jha <premjha2@in.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Greg Still <stillgs@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/135184
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>