Common routine for verifying PPE image is running on correct PPE engine
The routine retrieves the PPE type and instance ID from the PPE's PIR
and verifies it against the image's targeted PPE type and instance ID.

The routine halts the processor if the check fails.

There's a TODO item to also create an error log. But I don't believe all
the necessary plumbing is in place yet to implement this. Let me know if
wrong.

I've put the routine into a new shared header,
./powmanlip/pminit_common.h. There may be other init actions that we can
put into this header that is also common between our engines. Let me
know.

Also, let me know if you'd like me to update the commit to include your
engine's init tasks to call the routine.

Also note that, completely unrelated, I've used this commit to clean up
the debugging of the OINKR issue in dce_irq_handlers.c.

Change-Id: I4f38b11eeb1c822fe57e97de7d3ddaeac0729f10
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/139628
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: DAVID DU <daviddu@us.ibm.com>
Reviewed-by: RANGANATHPRASAD G. BRAHMASAMUDRA <prasadbgr@in.ibm.com>
Reviewed-by: Greg Still <stillgs@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/139630
Reviewed-by: Team <hostboot@us.ibm.com>
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>