Print error message if python version cannot be detected (MCP9 toolkit)
On some platforms there is no "python" binary/symbolic-link. Only python2 and python3
binaries are available. We should not default to python2 mistakenly but print
a proper error message.

The user can create /usr/bin/python with:

  sudo update-alternatives --config python

This detection is needed in multiple places, so moving it out to a dedicated
function that can be used from there.

Change-Id: I68740d4afd6ffa8de646f790105fe29200f26612
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/139782
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Rajees P P <rajerpp1@in.ibm.com>
Reviewed-by: Akhilesh S <akhis023@in.ibm.com>
Reviewed-by: AMIT J TENDOLKAR <amit.tendolkar@in.ibm.com>