WOF Data Collector (WDC) - GPE0 acting OCC (formerly on PGPE)
Reconfigured the OCC files, and necessary image build files, that
formerly were associated with the PGPE image to [temporarily] make
the PGPE perform the PSPI "drivers" to handle collection of sensor
data and IDDQ results we have planned for the OCC, namely sending
DB0 to DCE and accepting VC1/3 from DCE and OCE.

All these files have now been put into the ./wdc directory, which
is a sibling to all the other image dirs except it's in ./internal
rather than ./public.

The purpose of the wdc image is to be used in a future test
environment where we can test certain OCC or WOF functions but
without having to rely on the OCC image/code.

Since the modified IRQ target table in occhw_irq_config.h affects
the programmed content of the OCC's (Spinal's) interrupt routine
regs, OIRR0/1, whenever the WDC is engaged for testing purposes,
all other images on the Spinal must use the same modified IRQ
target table as otherwise the images, during initial Iota boot,
will end up writing different routes into the OIRR regs and which
might eventual be detected, depending on image loading order,
during boot of the images.  For this reason, the flag that
indicates WDC is active, namely WDC_ACTING_OCC, must be set by
all Spinal images. Thus, the introduction of the pmbuild_common.mk
build file.

Also added an error trace out in case the programming of the
IRR routing regs goes wrong.

Also added a default IRQ handler for all our *unused* but owned
and configured ext IRQs but that are not intended to be used in
interrupt context by the image. This is to prevent hanging the
PPE in case we should get interrupted by one of these IRQs, in
spite of the fact that we initially masked it. This could happen
in case we, or another proc sharing the same OCB, accidentally
unmasks the IRQ.

Change-Id: Idca2bd141cd27e1dc107e30d1ec3eb58299bd6fa
Original-Change-Id: I73b456919892df7c0874116ebec408c5b9d1dd73
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140764
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Prem Shanker Jha <premjha2@in.ibm.com>
Reviewed-by: Greg Still <stillgs@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/142009
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>