Code update: Base chip-op framework
This commit is to get the base chip-op framework
for code-update in place. The intention is to get
command-class and op-code for all chip-ops defined
in a source file and each called chip-op can then
be defined in separate files stacked on top of this
commit.

Also all the common functions that may be needed
across different code-update chip-ops can be defined
here.

- Added getCodeLevels chip-op (class AE, opcode 01)
- Added updateImage chip-op   (class AE, opcode 02)
- Added syncPartition chip-op (class AE, opcode 03)
- Added getCapability chip-op (class A8, opcode 02)

Jira Id: PFSBE-5

Change-Id: Iaa9b98e17c01ba6f907536a1a603232c2670d0c8
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140856
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sunil Kumar <skumar8j@in.ibm.com>
Reviewed-by: Ramesh I <rameshi1@in.ibm.com>
Reviewed-by: AMIT J TENDOLKAR <amit.tendolkar@in.ibm.com>