Create heap class, pak stack, pak load chipop
Creation of the Heap class, which will be added as a global variable
during pkboot. In main.C, this global heap variable will be initialized
with the calculated scratch_start pibmem address.

The heap class will manage a partitioned dynamic memory space, located
after the sbe/sppe compiled img in pibmem. The lowest heap address in
pibmem, growing to higher addresses will be the pakstack space. Any
new paks pushed into the pakstack will be added to this stack, moving
the heap_midline variable, which keeps track of the size of the pakstack
and the available space between the end of the packstack and the highest
address alloted to the heap space. Any allocated scratch space will
begin at the highest address alloted to the heap space, growing to lower
addresses. The scratch limit is defined by the heap_midline based on how
much memory the pakstack is using.

The pakLoad chipop will use the global heap object to manage pushing in
new paks from the fifo. The heap object will tell the chipop where to
begin writing the new pak and will update it's heap_midline variable
after as necessary. After each push, the current address is added to a
pak_ends array and pak_count is incremented. This is what the
popPakChipOp will use to find the end of the previous stack and write a
PAK_END marker, effectively telling any FileArchive object to not walk
the newly popped pak.

Change-Id: I73230d94fc8ca4afb09e79149529060bdc6de6bf
Signed-off-by: akhis023 <akhis023@in.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/141068
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sunil Kumar <skumar8j@in.ibm.com>
Reviewed-by: AMIT J TENDOLKAR <amit.tendolkar@in.ibm.com>