Wrapper to execute draminit HWP.
Jira: PFSBE-121

ody_sppe_draminint HWP will log errros and these error have to be
stored in the pibmem. This needs some change in the stream class
for writing the data in the pibmem.

PFSBE-271 will be used to calculate and verify the hash of the
binary () Synopsis and IOPPE) loaded from NOR by runtime image.

Change-Id: Ifacd17e783829ba8c700bd9a04f3c1d261e24a75
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/144555
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: AMIT J TENDOLKAR <amit.tendolkar@in.ibm.com>
Reviewed-by: Akhilesh S <akhis023@in.ibm.com>