pk_initialize: Fix junk in trace header image name
In APP_DEFINED_TRACE_BUFFER mode the image name string is copied from
read-only data. The source of that copy was not fixed size but the size
of the copy was, so if the string was shorter than the maximum we might
(and did) end up copying other rodata which the compiler placed after the
string along with the string itself, resulting in garbage in the image
name field.

Fix this by explicitly making the source string a fixed size.

Change-Id: Idf8feb2c9f59019f338b8e4cacf63cb71635345b
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/146235
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Doug Gilbert <dgilbert@us.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/146236
Reviewed-by: Team <hostboot@us.ibm.com>
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>