Initial move to common internal chipops
Currently, only moving the sbecmdbist.C/H files and setting up the
infrastructure for an internal common source for chipop code.

This change includes the sbeBist chipop in the following builds:
- odysseylab pnor
- odysseylab sppe
- metis sppe
- metis sbe

This change excludes the sbeBist chipop in the following builds:
- odyssey pnor
- metis pme

To achieve this divide of a separated chipop code source for
odysseylab builds, changes to `chipoptable.py` were required
to add a secondary `chipop_headers.H` sourced from the
odysseylab interal directory. A new flag has been added to
the args of `chipop_headers.H` to add this secondary include
when creating `chipop_table.C`.

Change-Id: I7d798f36f2c81f348ea415cddae436326fabc2e6
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/147653
Reviewed-by: Akhilesh S <akhis023@in.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Anuj Tiwari <anuj.tiwari2@ibm.com>
Reviewed-by: Daniel Kiss <kiss@de.ibm.com>
Reviewed-by: AMIT J TENDOLKAR <amit.tendolkar@in.ibm.com>