Multi-tap support and image-level check of PSPI link enable
This commit is in support of the multi-tap support we need
in the 405, and WDC, on the spinal side.

Since this support is via PSPI, this commit also introduces
a common API function to check PSPI enablement and possibly
initialize the PSPI links, or to verify them, all depending
on the build environment.

Minor error reporting have been introduced in the PSPI API.
Some precautions have been made in the API so that it can
be used on the OCC side as well.

Updates have been made to the WDC to duplicate, as much as
possible, the updates made for the 405 to support multi-
taps. This includes
- configuring VC buffer arrays, and
- checking and clearing OISR and VC tap summary reg.

Further, traceouts during 'ekb build' have been updated to
highlight if the images are being built for a non-product
purpose when engaging either BOOT_STANDALONE or
WDC_ACTING_OCC. All images report this context.

Lastly, various changes have been made to the images to make
their PSPI inits look and feel alike.

Key_Build=SPLIT_DIR_CHECK_IGNORE

Change-Id: I64cdc5c8df8afe27bac76c4d77374779d2fdb99b
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/147956
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: RANGANATHPRASAD G. BRAHMASAMUDRA <prasadbgr@in.ibm.com>
Reviewed-by: DAVID DU <daviddu@us.ibm.com>
Reviewed-by: Greg Still <stillgs@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/147960
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>