Move CBS cleanup steps into poz_tp_init
After the CBS has been started no outside code may touch any of the mailbox
registers any more since it could (and will) clash with code running on the
SPPE.

Actual lab fail case: The cbs_cleanup code (outside) was read-modify-writing
ROOT_CTRL0 while the SPPE was executing the command table. This caused the
outside code to restore the value of the PCB_RESET bit (cleared by command
table code) to 1, causing SCOMs to fail and the SPPE to crash.

The fix: Move the cbs_cleanup steps to either _before_ the CBS is triggered,
or to a place where it runs either inside the SPPE or the SPPE is quiesced
if it runs in Cronus - namely poz_tp_init.

Change-Id: I406216c2f4727de88073dcc2639999f141078679
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/150112
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: RYAN P KING <rpking@us.ibm.com>
Reviewed-by: Bryan Bartosh <bbartosh@ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/150113
Reviewed-by: Team <hostboot@us.ibm.com>
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>