Track scratch register usage in spreadsheet, with tool to generate Markdown
Introduce a spreadsheet, poz_scratch_regs.xlsx, which describes scratch
register usage across P11, Odyssey and zMetis. Also add a tool,
scratch_regs.py, which ingests this spreadsheet and generates text based
on a Jinja template.

The first application of this is to generate a Markdown description of
the scratch registers, but it can (and should) also be used to generate code
which reads and writes scratch registers and translates attribute values
back and forth.

Change-Id: I1b8ea6697eb6e327436a38892dbdd5943a6623ba
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/150349
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel Kiss <kiss@de.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/150351
Reviewed-by: Team <hostboot@us.ibm.com>
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>