Updated to check for row=0xFF, col=0xFF, hgt=0xFF for scalar attributes.
The code currently checks if the row=0x0, col=0x0, hgt=0x0
to update the scalar attributes. In case of array attributes,
the co-ordinates shall be row=0xFF, col=0xFF, hgt=0xFF for
update of entire array or the exact index if a specific element
needs to be updated. As the users will always update the entire
array, a design change has been made to specify row=0xFF,
col=0xFF, hgt=0xFF. This way, if the update attribute packet
is created by the user, they don't need to know if an attribute
is a scalar or an array type.

Change-Id: I29f66a125aa3d4a0105ae38022f44362a931261d
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/150506
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Rajees P P <rajerpp1@in.ibm.com>