Let 'addCopyright' script find correct LICENSE_PROLOG
This repo has different licenses for public and internal code. The
'addCopyright' script will look for 'LICENSE_PROLOG' in the directory
containing the source file or will iterate backwards through parent
directories until a 'LICENSE_PROLOG' file is found. So I moved the
Apache 2.0 license file to the 'public/' subdirectory and left the IBM
Confidential license file in the repo root directory. Then removed the
custom edits in the 'addCopyright' for the internal prolog that are no
longer needed.

Change-Id: Ieb344d3e31038dfa75757155a8242336935cb186
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/150622
Reviewed-by: AMIT J TENDOLKAR <amit.tendolkar@in.ibm.com>
Reviewed-by: Akhilesh S <akhis023@in.ibm.com>
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>