Update target_info struct to make sure PMIC and DT are a pair for Ody
The corresponding PMICs and DTs are to be considered a pair (PMIC0 and
DT are a pair and so and so forth). If either the PMIC or DT are not
present in a pair, considered that pair as not present and dont use it.

Key_Build=SPLIT_DIR_CHECK_IGNORE

ZEN:MST-2083
Change-Id: I3e000145c1186dce112f6841b627bf7bd63791d1
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/150680
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Dev-Ready: STEPHEN GLANCY <sglancy@us.ibm.com>
Reviewed-by: STEPHEN GLANCY <sglancy@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/150682
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>