archive: Force 8-byte alignment for FileArchive
This is a clumsy workaround for a PPE compiler quirk: When iv_firstFile
and iv_archiveLimit are both used close to each other, the PPE compiler
likes to load them both in one fell swoop using an lvd instruction. Nice
idea, but if the fields aren't properly aligned that causes a machine check.

Instead of convincing ppe-gcc to not emit an lvd in this case, make sure
that instances of FileArchive are aligned on an 8-byte boundary to work
around the problem.

Change-Id: Ie998ef8714be559c7224d7cd0bfef17f41b30cad
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/151598
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel Kiss <kiss@de.ibm.com>
Reviewed-by: Doug Gilbert <dgilbert@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/151599
Reviewed-by: Sunil Kumar <skumar8j@in.ibm.com>