scratch: Allow out-of-order freeing and support persistent blocks
Upon freeing a memory block, mark it as freed. Move the scratch_bottom
pointer only if and as long as the bottommost block is marked freed.

This way, scratch_free can be called in any order; if a block that is
not the most recently allocated one is freed, it is marked for cleanup
but its memory does not immediately become available again. Only when
there is a consecutive row of freed blocks starting at the bottommost
block, the entire chain will be cleaned up at once.

Along the way, implement a way to mark allocations as "persistent", which
makes them immune to scratch_free_all().

Change-Id: I7b696b1b766b3ccdfcf4467485eac7de716a54b9
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/151888
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sunil Kumar <skumar8j@in.ibm.com>