build: Introduce library concept, add "crypto" library
Introduce a way for modules/exes to easily include function libraries which
provide a set of functions which will be statically linked into the module
or exe instead of provided via a shared library. This way commonly used
utility code can be picked from a shared location without complicated
makefile mechanics or shared library dependencies.

Libraries should go into their own subdirectory of public/common/libs; the
makefile should build a module named after the library which includes all the
compiled C files; only the static .a library for the module will be used. The
makefile must also specify two variables:

LIB_(library name)_DEPLIBS specifies the libraries to link if the library is
used; usually this will be only the name of the library in question but may
include additional dependencies of the library itself.

LIB_(library name)_INCDIRS specifies the include paths to be added for users
of the library. This is usually the library directory itself.

Users can easily add the library to their module/exe by using the
ADD_MODULE_LIBRARY and ADD_EXE_LIBRARY macros. They will make sure that the
static library is built and linked in, and add the library include dirs to the
module/exe include dirs.

As an example use case, create a "crypto" library which provides sha3, crc32
and IBM ECC code, and change the archive code to use the library instead of
its own sha3 implementation.

Change-Id: I9f05a5388e588c9eb6782c28a17c232340990209
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/152677
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/152679
Reviewed-by: Daniel Kiss <kiss@de.ibm.com>