targeting: Add functionality to set Functional state based on ATTR_PG
Add plat_updateFunctionalState which is supposed to be run after the
attr_setup istep. This code will go through the entire list of targets
again and update the functional flags based on ATTR_PG. For the actual
decision of whether a target is functional or not it defers to functions
which can be overridden by chip specific implementations.

Change-Id: Ieb6a894f6474f267afeb17ee47c1277fb39eac39
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/153037
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Rajees P P <rajerpp1@in.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>