Add SPI library
Add a new function library for SPI communication. Drivers are implemented
as classes that are based on abstract base classes for easy combining of
different low-level and higher-level drivers.

For starters there is a base SPI driver which supports generic SPI
transactions, and a SEEPROM driver which builds upon the base driver to
implement SEEPROM read and write access.

Also add a wrapper called spitool which provides a command-line frontend
to the driver as well as a usage example.

Change-Id: I9d73bffb17729a6a09c0083d1001ee08a55e466a
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/154438
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>
Reviewed-by: Rajees P P <rajerpp1@in.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/154439