ipl: Use fixed OPCG_WAIT for all chiplets
The PCB network is not synchronous so per-chiplet delay times are not
guaranteed and we need to use other mechanisms for synchronous clock
start/stop. Thus, set up all chiplets to use the same wait value, which
means we can set them up via multicast and don't need the chiplet delay
table.

Change-Id: Ic1a3befd8e2dcfde6daee810c5bf00e9ffe35ef5
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/154424
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nick Landi <nlandi@ibm.com>
Reviewed-by: Hagen Schmidt <hschmidt@de.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/154425
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>