libcrypto: Switch to gzip compatible CRC32, use void*
The CRC32 scheme used by gzip is standardized as RFC1952 and available
in various languages' standard libraries. In comparison the bzip2 CRC
scheme is not as common, so switch to the gzip scheme before anyone really
starts using this :)

Also make the input data void* so callers don't have to typecast.

Change-Id: I5db26b52e2be82932e987c07ed3e636d011ca494
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/155916
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel Kiss <kiss@de.ibm.com>
Reviewed-by: Andrew Singer <andrew.singer@ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/155921
Reviewed-by: Team <hostboot@us.ibm.com>
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>