poz_fastarray: Allow for longer data pipelines
If we skip empty space with ARY clocks off we have to be sure we turn
ARY back on early enough that by the time we expect data again we get
it. Some paths from the array to the capture latches are deeper than
five clock cycles so increase the limit to ten.

Change-Id: If0bc51af6395da332c06515da694a0d8f6bdb11f
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/156723
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/156724
Reviewed-by: Daniel Kiss <kiss@de.ibm.com>