Resolve header file collisions
Compiling all of the headers fails because some of the different chip
headers have the same definitions. This is one option to resolve
the multiple definition problem. This commit removes all
defines from the P11 headers that are already present in the POZ
headers, and adds the POZ header include where needed.

Change-Id: I05d714847de3beda069bd846feac8c9e910cc5ea
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/158180
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Ben Gass <bgass@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/158207
Reviewed-by: Team <hostboot@us.ibm.com>
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>