Generate attribute update blob for the given Odyssey chip
In case of Metis, Odyssey SPD will be parsed by Metis SBE. After that,
Metis SBE will call getAttrUpdBlobSize() to get the buffer size required
to hold the generated blob and then pass that buffer to genOverride()
to get the update blob. Then, this blob will be sent to the
concerned Odyssey chip.

JIRA: PFSBE-716
Change-Id: Ib21a1bb719b61c9bf332252852fe33fb16a52509
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/160064
Reviewed-by: Rajees P P <rajerpp1@in.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sunil Kumar <skumar8j@in.ibm.com>
Reviewed-by: Daniel Kiss <kiss@de.ibm.com>
Reviewed-by: AMIT J TENDOLKAR <amit.tendolkar@in.ibm.com>