FAPI2: Add loadEmbeddedFile flag to suppress "not found" messages
Since loadEmbeddedFile is all we have we sometimes need to call it with
file names that may or may not exist, effectively probing the repository
for the existence of a file (for example in poz_putRingBackend).

To prevent this from spamming traces with unnecessary and misleading
error messages, add a flag that tells the platform that we're poking around
and it shouldn't print any errors if a file is not found.

Change-Id: Ia2df48bfaaef549142e07e7316d8d6ac73274744
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb-src/+/163793
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Kahn C Evans <kahnevan@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Andrew Singer <andrew.singer@ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/163797
Reviewed-by: Team <hostboot@us.ibm.com>