spi_memory: Fix possible double-write of final block
In error situations, if write_end() was called after write_data() failed,
the code would write the final block again (which write_data() already
did in response to the error). Fix this by bailing early if we're no longer
in write_active state, but allow the caller to call write_end() once even
if we just recovered.

Change-Id: I8bb675d4caad7d2de032c8993e6be3f8960f65ce
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb-src/+/164494
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/164495
Reviewed-by: Team <hostboot@us.ibm.com>