Add macro for easy procedure declaration
Instead of having to type out the function signature twice all the time
(and inevitably making mistakes that cause horrible fails in Cronus)
declare a macro that takes care of prototype and pointer type both.

Convert generic perv procedure headers to use the new macro, and commit
a script that takes care of the bulk of the conversion for others to use.

Change-Id: Ic16f62f55c431f69a3a48af8d17b79bf370c03aa
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb-src/+/165023
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Joachim Fenkes <fenkes@de.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/165098
Reviewed-by: Team <hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>