Commit-Id size changed from 8 to 16 chars in info.txt
This is done to make commit-Id size consistent across EKB, SBE
and dump component. This change is needed part of to avoid spurious
odyssey code-update a reason (among others) is removal of 8-bytes commit-Id
encoded in the dump hdct.bin file that need to be streamed separately
now reading from the info.txt file along with the dump.

JIRA:PFSBE-863
Change-Id: Ie0247ae59a4ff58b1f54ee5cef140666b38e628e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/165951
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Akhilesh S <akhis023@in.ibm.com>
Reviewed-by: Rangaswamy H B . <rangaswamy.h.b@ibm.com>